requirements.txt
# requirements.txt
# CEWP Pipeline - Complete Dependencies
# Generated: 2024-12 (Audit-Verified)
# ============================================

# --- Core Data Science & ML ---
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
joblib>=1.1.0

# --- Geospatial ---
geopandas>=0.10.0
h3>=3.7.0
shapely>=1.8.0
pyproj>=3.2.0
rtree>=0.9.7
rasterio>=1.2.0
osmnx>=1.1.0

# --- Database ---
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0
pyarrow>=6.0.0

# --- Configuration ---
python-dotenv>=0.19.0
pyyaml>=5.4.0

# --- Cloud & APIs ---
earthengine-api>=0.1.300
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
db-dtypes>=1.0.0
yfinance>=0.2.0
requests>=2.26.0

# --- Utilities ---
tqdm>=4.60.0
tenacity>=8.0.0

# --- Visualization (Optional but recommended) ---
matplotlib>=3.4.0
seaborn>=0.11.0

# ============================================
# SYSTEM REQUIREMENTS (Not pip-installable):
# ============================================
# 1. PostgreSQL 14+ with extensions:
#    - PostGIS
#    - h3 (pgxn install h3)
#    - h3_postgis
#
# 2. GDAL/OGR libraries (for geopandas/rasterio):
#    - Windows: Use conda or OSGeo4W
#    - Linux: apt install gdal-bin libgdal-dev
#    - macOS: brew install gdal
#
# 3. Google Cloud SDK (for BigQuery/GEE auth):
#    - gcloud auth application-default login
#
# ============================================
# ENVIRONMENT VARIABLES REQUIRED (.env file):
# ============================================
# DB_USER=postgres
# DB_PASS=your_password
# DB_HOST=localhost
# DB_PORT=5432
# DB_NAME=car_cewp
# ACLED_API_KEY=your_key
# IOM_PRIMARY_KEY=your_key
# FEWS_NET_EMAIL=your_email
# FEWS_NET_PASSWORD=your_password
# ============================================
